refactor(theme): co-locate ThemeProvider under theme/ + body defaults via CSS
Address review feedback:
- Move ThemeProvider next to ThemeContext under src/theme/ (context and
  provider co-located; public export path via providers/index.ts unchanged).
- Replace the imperative document.body inline-style effect with a declarative
  `body { color; background-color }` rule in src/theme/styles/global.css using
  the token custom properties. It swaps with `data-cui-theme` and ships in the
  `@layer clickui` layer, so any consumer `body` rule overrides it. Non-breaking:
  ClickUIProvider sets data-cui-theme, and this is lower-priority than the old
  inline styles (which would have clobbered a consumer's own body color).
